requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
itsdangerous==2.1.2
Werkzeug==3.0.1
Jinja2==3.1.2
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.8
requests==2.31.0
line-bot-sdk==3.7.0
